Use ruby refinements to monkey-patch string
