+++ lib/business_central/object/object_helper.rb
@@ -3,9 +3,11 @@
 module BusinessCentral
   module Object
     module ObjectHelper
+      using Refinements::Strings
+
       def method_missing(object_name, **params)
-        if BusinessCentral::Object.const_defined?(object_name.to_s.classify)
-          klass = BusinessCentral::Object.const_get(object_name.to_s.classify)
+        if BusinessCentral::Object.const_defined?(object_name.to_s.to_class_sym)
+          klass = BusinessCentral::Object.const_get(object_name.to_s.to_class_sym)
           klass.new(self, **params)
         else
           BusinessCentral::Object::Base.new(self, **params.merge!({ object_name: object_name }))
