+++ lib/business_central/object/picture.rb
@@ -3,6 +3,8 @@
 module BusinessCentral
   module Object
     class Picture < Base
+      using Refinements::Strings
+
       def update(id, data)
         object = find_all
         Request.call(:patch, @client, "#{build_url(object_id: id)}/content", etag: object[:etag], params: {}) do |request|
