+++ lib/business_central/object/url_builder.rb
@@ -3,6 +3,8 @@
 module BusinessCentral
   module Object
     class URLBuilder
+      using Refinements::Strings
+
       extend URLHelper
 
       class << self
